+++ therapi-runtime-javadoc/src/main/java/com/github/therapi/runtimejavadoc/FieldJavadoc.java
@@ -8,6 +8,15 @@ public class FieldJavadoc extends BaseJavadoc {
         super(name, comment, seeAlso, other);
     }
 
+    public static FieldJavadoc createEmpty(String fieldName) {
+        return new FieldJavadoc(fieldName, null, null, null) {
+            @Override
+            public boolean isEmpty() {
+                return true;
+            }
+        };
+    }
+
     @Override
     public String toString() {
         return "FieldJavadoc{"