+++ therapi-runtime-javadoc/src/main/java/com/github/therapi/runtimejavadoc/RuntimeJavadoc.java
@@ -1,23 +1,22 @@
 package com.github.therapi.runtimejavadoc;
 
+import com.eclipsesource.json.Json;
+import com.eclipsesource.json.JsonObject;
+import com.github.therapi.runtimejavadoc.internal.JsonJavadocReader;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.List;
-import java.util.Optional;
-
-import com.eclipsesource.json.Json;
-import com.eclipsesource.json.JsonObject;
-import com.github.therapi.runtimejavadoc.internal.JsonJavadocReader;
 
 import static com.github.therapi.runtimejavadoc.internal.RuntimeJavadocHelper.javadocResourceSuffix;
 import static java.nio.charset.StandardCharsets.UTF_8;
 
 /**
- * Allows access to Javadoc elements at runtime. This will only find the Javadoc of elements that were processed by
- * the annotation processor.
+ * Allows access to Javadoc elements at runtime for code that was compiled using the
+ * {@code therapi-runtime-javadoc-scribe} annotation processor.
  */
 public class RuntimeJavadoc {
 
@@ -27,39 +26,41 @@ public class RuntimeJavadoc {
 
     /**
      * Gets the Javadoc of the given class.
+     * <p>
+     * The return value is always non-null. If no Javadoc is available, the returned object's
+     * {@link BaseJavadoc#isEmpty isEmpty()} method will return {@code true}.
      *
-     * @param clazz
-     *         the class to retrieve the Javadoc for
-     *
-     * @return the Javadoc of the given class, or an empty optional if no documentation was found
+     * @param clazz the class whose Javadoc you want to retrieve
+     * @return the Javadoc of the given class
      */
-    public static Optional<ClassJavadoc> getJavadoc(Class clazz) {
+    public static ClassJavadoc getJavadoc(Class clazz) {
         return getJavadoc(clazz.getName(), clazz);
     }
 
     /**
      * Gets the Javadoc of the given class.
+     * <p>
+     * The return value is always non-null. If no Javadoc is available, the returned object's
+     * {@link BaseJavadoc#isEmpty isEmpty()} method will return {@code true}.
      *
-     * @param qualifiedClassName
-     *         the fully qualified name of the class to retrieve the Javadoc for
-     *
-     * @return the Javadoc of the given class, or an empty optional if no documentation was found
+     * @param qualifiedClassName the fully qualified name of the class whose Javadoc you want to retrieve
+     * @return the Javadoc of the given class
      */
-    public static Optional<ClassJavadoc> getJavadoc(String qualifiedClassName) {
+    public static ClassJavadoc getJavadoc(String qualifiedClassName) {
         return getJavadoc(qualifiedClassName, RuntimeJavadoc.class);
     }
 
     /**
      * Gets the Javadoc of the given class, using the given {@link ClassLoader} to find the Javadoc resource.
+     * <p>
+     * The return value is always non-null. If no Javadoc is available, the returned object's
+     * {@link BaseJavadoc#isEmpty isEmpty()} method will return {@code true}.
      *
-     * @param qualifiedClassName
-     *         the fully qualified name of the class to retrieve the Javadoc for
-     * @param classLoader
-     *         the class loader to use to find the Javadoc resource file
-     *
-     * @return the Javadoc of the given class, or an empty optional if no documentation was found
+     * @param qualifiedClassName the fully qualified name of the class whose Javadoc you want to retrieve
+     * @param classLoader        the class loader to use to find the Javadoc resource file
+     * @return the Javadoc of the given class
      */
-    public static Optional<ClassJavadoc> getJavadoc(String qualifiedClassName, ClassLoader classLoader) {
+    public static ClassJavadoc getJavadoc(String qualifiedClassName, ClassLoader classLoader) {
         final String resourceName = getResourceName(qualifiedClassName);
         try (InputStream is = classLoader.getResourceAsStream(resourceName)) {
             return parseJavadocResource(qualifiedClassName, is);
@@ -70,15 +71,15 @@ public class RuntimeJavadoc {
 
     /**
      * Gets the Javadoc of the given class, using the given {@link Class} object to load the Javadoc resource.
+     * <p>
+     * The return value is always non-null. If no Javadoc is available, the returned object's
+     * {@link BaseJavadoc#isEmpty isEmpty()} method will return {@code true}.
      *
-     * @param qualifiedClassName
-     *         the fully qualified name of the class to retrieve the Javadoc for
-     * @param loader
-     *         the class object to use to find the Javadoc resource file
-     *
-     * @return the Javadoc of the given class, or an empty optional if no documentation was found
+     * @param qualifiedClassName the fully qualified name of the class whose Javadoc you want to retrieve
+     * @param loader             the class object to use to find the Javadoc resource file
+     * @return the Javadoc of the given class
      */
-    public static Optional<ClassJavadoc> getJavadoc(String qualifiedClassName, Class loader) {
+    public static ClassJavadoc getJavadoc(String qualifiedClassName, Class loader) {
         final String resourceName = getResourceName(qualifiedClassName);
         try (InputStream is = loader.getResourceAsStream("/" + resourceName)) {
             return parseJavadocResource(qualifiedClassName, is);
@@ -91,9 +92,9 @@ public class RuntimeJavadoc {
         return qualifiedClassName.replace(".", "/") + javadocResourceSuffix();
     }
 
-    private static Optional<ClassJavadoc> parseJavadocResource(String qualifiedClassName, InputStream is) throws IOException {
+    private static ClassJavadoc parseJavadocResource(String qualifiedClassName, InputStream is) throws IOException {
         if (is == null) {
-            return Optional.empty();
+            return ClassJavadoc.createEmpty(qualifiedClassName);
         }
 
         try (InputStreamReader r = new InputStreamReader(is, UTF_8)) {
@@ -105,66 +106,75 @@ public class RuntimeJavadoc {
     /**
      * Gets the Javadoc of the given method.
      * <p>
+     * The return value is always non-null. If no Javadoc is available, the returned object's
+     * {@link BaseJavadoc#isEmpty isEmpty()} method will return {@code true}.
+     * <p>
      * Implementation note: this method first retrieves the Javadoc of the class, and then matches the method signature
      * with the correct documentation. If the client code's purpose is to loop through all methods doc, prefer using
      * {@link #getJavadoc(Class)} (or one of its overloads), and calling {@link ClassJavadoc#getMethods()} on the
      * returned class doc to retrieve method docs.
      *
-     * @param method
-     *         the method to get the Javadoc for
-     *
-     * @return the given method's Javadoc, or an empty optional if no documentation was found
+     * @param method the method whose Javadoc you want to retrieve
+     * @return the given method's Javadoc
      */
-    public static Optional<MethodJavadoc> getJavadoc(Method method) {
-        Optional<ClassJavadoc> javadoc = getJavadoc(method.getDeclaringClass());
-        return javadoc.map(ClassJavadoc::getMethods).flatMap(mDocs -> findMethodJavadoc(mDocs, method));
+    public static MethodJavadoc getJavadoc(Method method) {
+        ClassJavadoc javadoc = getJavadoc(method.getDeclaringClass());
+        return findMethodJavadoc(javadoc.getMethods(), method);
     }
 
-    private static Optional<MethodJavadoc> findMethodJavadoc(List<MethodJavadoc> methodDocs, Method method) {
-        return methodDocs.stream().filter(m -> m.matches(method)).findAny();
+    private static MethodJavadoc findMethodJavadoc(List<MethodJavadoc> methodDocs, Method method) {
+        for (MethodJavadoc methodJavadoc : methodDocs) {
+            if (methodJavadoc.matches(method)) {
+                return methodJavadoc;
+            }
+        }
+        return MethodJavadoc.createEmpty(method);
     }
 
     /**
      * Gets the Javadoc of the given field.
+     * <p>
+     * The return value is always non-null. If no Javadoc is available, the returned object's
+     * {@link BaseJavadoc#isEmpty isEmpty()} method will return {@code true}.
      * <p>
      * Implementation note: this method first retrieves the Javadoc of the class, and then matches the field name
      * with the correct documentation. If the client code's purpose is to loop through all fields doc, prefer using
      * {@link #getJavadoc(Class)} (or one of its overloads), and calling {@link ClassJavadoc#getFields()} on the
      * returned class doc to retrieve field docs.
      *
-     * @param field
-     *         the field to get the Javadoc for
-     *
-     * @return the given field's Javadoc, or an empty optional if no documentation was found
+     * @param field the field whose Javadoc you want to retrieve
+     * @return the given field's Javadoc
      */
-    public static Optional<FieldJavadoc> getJavadoc(Field field) {
-        Optional<ClassJavadoc> javadoc = getJavadoc(field.getDeclaringClass());
-        return javadoc.map(ClassJavadoc::getFields).flatMap(fDocs -> findFieldJavadoc(fDocs, field));
-    }
-
-    private static Optional<FieldJavadoc> findFieldJavadoc(List<FieldJavadoc> fieldDocs, Field field) {
-        return fieldDocs.stream().filter(m -> m.getName().equals(field.getName())).findAny();
+    public static FieldJavadoc getJavadoc(Field field) {
+        ClassJavadoc javadoc = getJavadoc(field.getDeclaringClass());
+        return findFieldJavadoc(javadoc.getFields(), field.getName());
     }
 
     /**
      * Gets the Javadoc of the given enum constant.
+     * <p>
+     * The return value is always non-null. If no Javadoc is available, the returned object's
+     * {@link BaseJavadoc#isEmpty isEmpty()} method will return {@code true}.
      * <p>
      * Implementation note: this method first retrieves the Javadoc of the class, and then matches the enum constant's
      * name with the correct documentation. If the client code's purpose is to loop through all enum constants docs,
      * prefer using {@link #getJavadoc(Class)} (or one of its overloads), and calling
      * {@link ClassJavadoc#getEnumConstants()} on the returned class doc to retrieve enum constant docs.
      *
-     * @param enumValue
-     *         the enum constant to get the Javadoc for
-     *
-     * @return the given enum constant's Javadoc, or an empty optional if no documentation was found
+     * @param enumValue the enum constant whose Javadoc you want to retrieve
+     * @return the given enum constant's Javadoc
      */
-    public static Optional<FieldJavadoc> getJavadoc(Enum<?> enumValue) {
-        Optional<ClassJavadoc> javadoc = getJavadoc(enumValue.getDeclaringClass());
-        return javadoc.map(ClassJavadoc::getEnumConstants).flatMap(fDocs -> findEnumValueJavadoc(fDocs, enumValue));
+    public static FieldJavadoc getJavadoc(Enum<?> enumValue) {
+        ClassJavadoc javadoc = getJavadoc(enumValue.getDeclaringClass());
+        return findFieldJavadoc(javadoc.getEnumConstants(), enumValue.name());
     }
 
-    private static Optional<FieldJavadoc> findEnumValueJavadoc(List<FieldJavadoc> fieldDocs, Enum<?> enumValue) {
-        return fieldDocs.stream().filter(m -> m.getName().equals(enumValue.name())).findAny();
+    private static FieldJavadoc findFieldJavadoc(List<FieldJavadoc> fieldDocs, String fieldName) {
+        for (FieldJavadoc fDoc : fieldDocs) {
+            if (fDoc.getName().equals(fieldName)) {
+                return fDoc;
+            }
+        }
+        return FieldJavadoc.createEmpty(fieldName);
     }
 }